samples: net: sockets: Add CoAP service example
Replaced the CoAP server example with CoAP services.
