+++ samples/net/sockets/coap_server/CMakeLists.txt
@@ -7,3 +7,5 @@ project(coap_server)
 FILE(GLOB app_sources src/*.c)
 target_sources(app PRIVATE ${app_sources})
 target_include_directories(app PRIVATE ${ZEPHYR_BASE}/subsys/net/ip)
+
+zephyr_linker_sources(DATA_SECTIONS sections-ram.ld)
