+++ samples/net/sockets/coap_server/README.rst
@@ -1,13 +1,18 @@
 .. zephyr:code-sample:: coap-server
-   :name: CoAP server
-   :relevant-api: coap udp
+   :name: CoAP service
+   :relevant-api: coap coap_service udp
 
-   Use the CoAP library to implement a server that exposes CoAP resources.
+   Use the CoAP server subsystem to register CoAP resources.
 
 Overview
 ********
 
-This sample is a simple CoAP server showing how to expose a simple resource.
+This sample shows how to register CoAP resources to a main CoAP service.
+The CoAP server implementation expects all services and resources to be
+available at compile time, as they are put into dedicated sections.
+
+The resource is placed into the correct linker section based on the owning
+service's name. A linker file is required, see ``sections-ram.ld`` for an example.
 
 This demo assumes that the platform of choice has networking support,
 some adjustments to the configuration may be needed.
