+++ samples/net/sockets/coap_server/boards/native_posix.conf
@@ -1,0 +1,3 @@
+CONFIG_NET_L2_ETHERNET=y
+CONFIG_NATIVE_POSIX_SLOWDOWN_TO_REAL_TIME=y
+CONFIG_NATIVE_UART_0_ON_STDINOUT=y
