+++ samples/net/sockets/coap_server/boards/qemu_cortex_m3.conf
@@ -1,0 +1,4 @@
+CONFIG_NET_L2_ETHERNET=y
+CONFIG_ETH_DRIVER=y
+CONFIG_ETH_STELLARIS=y
+CONFIG_NET_QEMU_ETHERNET=y
