+++ samples/net/sockets/coap_server/prj.conf
@@ -10,7 +10,10 @@ CONFIG_NET_SOCKETS_POLL_MAX=4
 
 # CoAP
 CONFIG_COAP=y
+CONFIG_COAP_SERVER=y
+CONFIG_COAP_SERVER_WELL_KNOWN_CORE=y
 CONFIG_COAP_WELL_KNOWN_BLOCK_WISE=n
+CONFIG_COAP_OBSERVER_EVENTS=y
 
 # Kernel options
 CONFIG_ENTROPY_GENERATOR=y
@@ -22,9 +25,7 @@ CONFIG_NET_LOG=y
 
 # Network Shell
 CONFIG_NET_SHELL=y
-
-CONFIG_MAIN_STACK_SIZE=2048
-CONFIG_HEAP_MEM_POOL_SIZE=4096
+CONFIG_COAP_SERVER_SHELL=y
 
 # Configuration
 CONFIG_NET_CONFIG_SETTINGS=y
