+++ samples/net/sockets/coap_server/sample.yaml
@@ -1,12 +1,14 @@
 common:
   filter: CONFIG_FULL_LIBC_SUPPORTED and not CONFIG_NATIVE_LIBC
 sample:
-  description: TBD
-  name: TBD
+  description: BSD Sockets API CoAP server example
+  name: socket_coap_server
 tests:
   sample.net.sockets.coap_server:
     harness: net
     tags:
       - net
       - socket
-    platform_allow: qemu_x86
+    platform_allow:
+      - native_posix
+      - qemu_x86
