+++ samples/net/sockets/coap_server/sections-ram.ld
@@ -1,0 +1,5 @@
+/* SPDX-License-Identifier: Apache-2.0 */
+
+#include <zephyr/linker/iterable_sections.h>
+
+ITERABLE_SECTION_RAM(coap_resource_coap_server, 4)
